whole nums

+++ backend/db/migrations/20240923152153-create-spot.js
@@ -55,7 +55,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.DECIMAL(30, 2),
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       createdAt: {
